+++ runner/1920-1080.json
@@ -7,7 +7,7 @@
       "windows": [
         {
           "border": false,
-          "name": "Image Test",
+          "name": "RegressionTest",
           "pos": { "x": 0, "y": 0 },
           "size": { "x": 960, "y": 540 },
           "res": { "x": 1920, "y": 1080 },
@@ -31,14 +31,10 @@
       ]
     }
   ],
-  "settings": {
-    "display": {
-      "swapinterval": 1
-    }
-  },
+  "settings": { "display": { "swapinterval": 1 } },
   "users": [
     {
-      "eyeseparation": 0.06499999761581421,
+      "eyeseparation": 0.065,
       "pos": { "x": 0.0, "y": 0.0, "z": 0.0 }
     }
   ],
