+++ runner/testsuite/constants.py
@@ -1,0 +1,1 @@
+test_base_dir = "tests/visual"
