+++ runner/testsuite/openspace.py
@@ -1,0 +1,190 @@
+##########################################################################################
+#                                                                                        #
+# OpenSpace Visual Testing                                                               #
+#                                                                                        #
+# Copyright (c) 2024                                                                     #
+#                                                                                        #
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this   #
+# software and associated documentation files (the "Software"), to deal in the Software  #
+# without restriction, including without limitation the rights to use, copy, modify,     #
+# merge, publish, distribute, sublicense, and/or sell copies of the Software, and to     #
+# permit persons to whom the Software is furnished to do so, subject to the following    #
+# conditions:                                                                            #
+#                                                                                        #
+# The above copyright notice and this permission notice shall be included in all copies  #
+# or substantial portions of the Software.                                               #
+#                                                                                        #
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,    #
+# INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A          #
+# PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT     #
+# HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF   #
+# CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE   #
+# OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                          #
+##########################################################################################
+
+import asyncio
+import glob
+import os
+import subprocess
+import time
+from openspace import Api
+from .test import Test, TestResult
+
+
+
+def write_configuration_overwrite(base_path, data_path):
+  """
+  Creates a openspace.cfg override file that sets up a common testing environment. These
+  are mostly for enabling caching to reduce the amount of testing and removing as much of
+  the logging as possible which would otherwise pollute the usage results
+  """
+  with open(f"{base_path}/openspace.cfg.override", "w") as f:
+    # Use a common sync folder outside of the build to prevent redownloading of data
+    sync_location = f"{data_path}/sync"
+    f.write(f"Paths.SYNC = [[{sync_location}]]\n")
+
+    # Enable MRF caching for the same reason and to reduce dependency on external servers
+    mrf_location = f"{data_path}/mrf"
+    f.write("ModuleConfigurations.GlobeBrowsing.MRFCacheEnabled = true\n")
+    f.write(f"ModuleConfigurations.GlobeBrowsing.MRFCacheLocation = [[{mrf_location}]]\n")
+
+    # Disable the UI as it contains different time-dependent content
+    f.write(f"ModuleConfigurations.CefWebGui.Enabled = false\n")
+
+    # Remove the version checking URL as it would otherwise count as a "user"
+    f.write("VersionCheckUrl = [[]]\n")
+
+    # Checking the OpenGL state will cause OpenGL errors to appear in the log
+    f.write("CheckOpenGLState = true\n")
+
+    # We can reduce the amount of time that we have to wait for OpenSpace to shut down
+    f.write("ShutdownCountdown = 0.25\n")
+
+
+
+async def setup_test_run(openspace):
+  """
+  Setup settings that a common to all test runs. This are, in general, setting that are
+  reasonably different between runs of OpenSpace, such as the local time, the commit hash,
+  and others
+  """
+  # We always want to start paused to prevent some timing-related inconsistencies
+  await openspace.time.setPause(True)
+
+  # Unless explicitly added, we don't want display elements that show variable content
+  #  Dashboard: Framerate
+  #  ScreenLog: Log message retention
+  #  Version: Contains the commit hash
+  #  Camera: Not technically needed, but results in a cleaner time
+  await openspace.setPropertyValueSingle("Dashboard.IsEnabled", False)
+  await openspace.setPropertyValueSingle("RenderEngine.ShowLog", False)
+  await openspace.setPropertyValueSingle("RenderEngine.ShowVersion", False)
+  await openspace.setPropertyValueSingle("RenderEngine.ShowCamera", False)
+
+
+
+async def internal_run(openspace, test):
+  """
+  This function runs the actual test in the library object passed into it. It first sets
+  up default values, then runs the individual instructions for the test, and retreives
+  other information such as the screenshot location, and the commit hash from OpenSpace.
+
+  This function assumes that the `openspace` instance is already authenticated and
+  connected to the OpenSpace instance and is ready to take commands.
+  """
+  print(f"  Starting test")
+  await setup_test_run(openspace)
+  await test.run(openspace)
+  print("  Finished test")
+
+  # Give the screenshot writing some time to finish
+  time.sleep(1)
+
+  # Get the location of the screenshot folder from OpenSpace. It should always be the
+  # same but this is just to make sure it will work
+  screenshot_folder = await openspace.absPath("${SCREENSHOTS}")
+
+  # Get the commit hash from OpenSpace itself
+  version = await openspace.version()
+  commit = version["Commit"]
+
+  await openspace.toggleShutdown()
+
+  return (screenshot_folder, commit)
+
+
+
+def run_single_test(test_path, executable) -> TestResult:
+  """
+  Run the single test provided by `test_path` using the OpenSpace executable provided by
+  `executable`. This will include starting OpenSpace as subprocess using a known
+  configuration file and the profile required by the profile, establishing a connection
+  using the Python API to the OpenSpace instance, running the test's instructions and
+  returning the results.
+
+   - `test_path`: The path to the ostest file that should be run. This file must exist
+   - `executable`: The path to the OpenSpace executable that should be run for the tests
+  """
+  print(f"Running test: {test_path}")
+  test = Test(test_path)
+
+  start_time = time.perf_counter()
+  print("  Starting OpenSpace")
+  process = subprocess.Popen(
+    [
+      executable,
+      "--config", f"{os.getcwd()}/1920-1080.json",
+      "--profile", test.profile,
+      "--bypassLauncher"
+    ],
+    cwd=os.path.dirname(executable),
+    stdout=subprocess.DEVNULL,
+    stderr=subprocess.PIPE
+  )
+
+  # Add a sleeping time instead of repeatedly trying to reconnect. Starting up OpenSpace
+  # in general takes longer than this, so we don't actually lose any time
+  time.sleep(15)
+
+
+  async def mainLoop():
+    """
+    The main loop of an async event loop that is needed for the OpenSpace Python API to
+    work correctly. This will connect to OpenSpace, which then triggers the rest of the
+    test run
+    """
+    print("  Connecting...")
+    os_api = Api("localhost", 4681)
+    os_api.connect()
+    openspace = await os_api.singleReturnLibrary()
+    print("  Connected to OpenSpace")
+    # Run the current test
+    (screenshot_folder, commit) = await asyncio.create_task(internal_run(openspace, test))
+    os_api.disconnect()
+    return (screenshot_folder, commit)
+
+  (screenshot_folder, commit) = asyncio.new_event_loop().run_until_complete(mainLoop())
+
+
+  # Another wait while OpenSpace is shutting down
+  time.sleep(2)
+
+  # Get the error log from OpenSpace
+  error_log = process.stderr.read().decode()
+
+  # Kill OpenSpace
+  process.kill()
+  end_time = time.perf_counter()
+
+  # Collect all screenshots taken by the test
+  files = glob.glob(f"{screenshot_folder}/*.png")
+  print(f"Test images: {files}")
+
+  result = TestResult()
+  result.group = test.group
+  result.name = test.name
+  result.files = files
+  result.timing = end_time - start_time
+  result.commit = commit
+  result.error = error_log
+  return result
