+++ runner/testsuite/test.py
@@ -1,0 +1,103 @@
+##########################################################################################
+#                                                                                        #
+# OpenSpace Visual Testing                                                               #
+#                                                                                        #
+# Copyright (c) 2024                                                                     #
+#                                                                                        #
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this   #
+# software and associated documentation files (the "Software"), to deal in the Software  #
+# without restriction, including without limitation the rights to use, copy, modify,     #
+# merge, publish, distribute, sublicense, and/or sell copies of the Software, and to     #
+# permit persons to whom the Software is furnished to do so, subject to the following    #
+# conditions:                                                                            #
+#                                                                                        #
+# The above copyright notice and this permission notice shall be included in all copies  #
+# or substantial portions of the Software.                                               #
+#                                                                                        #
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,    #
+# INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A          #
+# PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT     #
+# HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF   #
+# CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE   #
+# OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                                          #
+##########################################################################################
+
+import json
+import os
+import time
+from .instruction import Instruction
+from .constants import test_base_dir
+
+class TestResult:
+  """
+  Stores the result of a single test run. It has the following members:
+    - `group`: The name of the group for which this is the test result
+    - `name`: The name of the test for which is the result
+    - `files`: The list of screenshots that were taken during the test. Each entry in this
+               list is a path to an image file
+    - `timing`: The number of seconds it took to execute the test
+    - `commit`: The commit hash for OpenSpace that was used to run the test
+    - `error`: The contents of the error stream that was captured during the test run
+  """
+  group: str
+  name: str
+  files = list[str]
+  timing: float
+  commit: str
+  error: str
+
+class Test:
+  """
+  This class represents an entire test run, consisting of multiple Instructions and a
+  profile that should be used.
+  """
+  def __init__(self, path: str):
+    assert(os.path.isfile(path))
+    self.test_path = path
+
+    with open(path) as f:
+      content = json.load(f)
+
+    if content["profile"] is None:
+      raise f"Missing 'profile' in test {path}'"
+    self.profile = content["profile"]
+
+    if content["commands"] is None:
+      raise Exception(f"Missing 'commands' in test {path}")
+
+    self.instructions = []
+    for command in content["commands"]:
+      try:
+        self.instructions.append(Instruction(command))
+      except Exception as error:
+        raise Exception(f"Error loading test {path}: {error}")
+
+    found_screenshot_instruction = False
+    for instruction in self.instructions:
+      if instruction.type == "screenshot":
+        found_screenshot_instruction = True
+        break
+
+    if not found_screenshot_instruction:
+      raise Exception(f"Error loading test {path}: No screenshot instruction")
+
+
+    # Get the testname by removing everything before (and including) "test/visual" and
+    # also removing the extension
+    start_idx = self.test_path.find(test_base_dir) + len(test_base_dir) + 1
+    full = self.test_path[start_idx:-len(".ostest")]
+    parts = full.split("/")
+
+    # The last part is the name of test, all others are combined to make the grouping
+    self.group = "-".join(parts[0:-1])
+    self.name = parts[-1]
+
+
+  async def run(self, openspace):
+    """
+    Runs the actual instructions on the provided OpenSpace API instance. There is a
+    mandatory 250ms wait time between every instructions
+    """
+    for instruction in self.instructions:
+      await instruction.run(openspace)
+      time.sleep(0.25)
